+++ libs/compiler/test/bench/compiler_bench_test.go
@@ -167,7 +167,7 @@ func BenchmarkCompileEmpty(b *testing.B) {
 
 	b.ResetTimer()
 	for i := 0; i < b.N; i++ {
-		_, _ = compiler.Compile(ctx, opts)
+		_ = compiler.Compile(ctx, opts)
 	}
 }